Fix migration code

+++ node-tools/src/main/scala/im/mak/nodetools/PayoutDBMigrate.scala
@@ -5,6 +5,8 @@ import com.wavesplatform.state.Blockchain
 import com.wavesplatform.transaction.transfer.MassTransferTransaction
 import com.wavesplatform.utils.ScorexLogging
 
+import scala.util.Try
+
 //noinspection TypeAnnotation
 object PayoutDBMigrate extends ScorexLogging {
   import PayoutDB._
@@ -29,46 +31,55 @@ object PayoutDBMigrate extends ScorexLogging {
   private[this] implicit val payoutsMeta            = schemaMeta[Payout]("payouts")
   private[this] implicit val payoutTransactionsMeta = schemaMeta[PayoutTransaction]("payout_transactions")
 
-  def migratePayouts(b: Blockchain): Unit = ctx.transaction {
-    val oldPayouts = run(quote(query[Payout].filter(p => query[PayoutTransaction].filter(_.payoutId == p.id).isEmpty)))
-    val withTxs = oldPayouts.map(
-      p =>
-        p.txId match {
-          case Some(txId) =>
-            val tx = b.transactionInfo(Base58.decode(txId)).map(_._2)
-            p -> tx
+  def migratePayouts(b: Blockchain): Unit = {
+    def migrationNeeded =
+      Try(executeQuery[ResultRow]("select tx_id from payouts")).isSuccess
 
-          case None =>
-            p -> None
-        }
-    )
+    if (!migrationNeeded) return
 
-    def addTransaction(id: Int, tx: MassTransferTransaction) = quote {
-      val txId = liftQ(tx.id().toString)
-      query[PayoutTransaction].insert(_.payoutId -> liftQ(id), _.id -> txId, _.transaction -> liftQ(tx))
+    ctx.transaction {
+      log.info("Starting DB migration")
+      val oldPayouts = run(quote(query[Payout].filter(p => query[PayoutTransaction].filter(_.payoutId == p.id).isEmpty)))
+      val withTxs = oldPayouts.map(
+        p =>
+          p.txId match {
+            case Some(txId) =>
+              val tx = b.transactionInfo(Base58.decode(txId)).map(_._2)
+              p -> tx
+
+            case None =>
+              p -> None
+          }
+      )
+
+      def addTransaction(id: Int, tx: MassTransferTransaction) = quote {
+        val txId = liftQ(tx.id().toString)
+        query[PayoutTransaction].insert(_.payoutId -> liftQ(id), _.id -> txId, _.transaction -> liftQ(tx))
+      }
+
+      def delPayout(id: Int) = quote {
+        query[Payout].filter(_.id == liftQ(id)).delete
+      }
+
+      withTxs.foreach {
+        case (p, tx) =>
+          tx match {
+            case Some(mtt: MassTransferTransaction) =>
+              log.info(s"Registering transaction for $p: $mtt")
+              run(addTransaction(p.id, mtt))
+            case Some(tx) =>
+              sys.error(s"Not mass transfer: $tx")
+            case None =>
+              log.warn(s"Transaction for $p not found, removing")
+              run(delPayout(p.id))
+          }
+      }
+
+      executeAction("""
+                      |alter table payouts
+                      |    drop column active_leases, tx_id, tx_height, confirmed;
+                      |""".stripMargin)
+      log.info("Migration finished")
     }
-
-    def delPayout(id: Int) = quote {
-      query[Payout].filter(_.id == liftQ(id)).delete
-    }
-
-    withTxs.foreach {
-      case (p, tx) =>
-        tx match {
-          case Some(mtt: MassTransferTransaction) =>
-            log.info(s"Registering transaction for $p: $mtt")
-            run(addTransaction(p.id, mtt))
-          case Some(tx) =>
-            sys.error(s"Not mass transfer: $tx")
-          case None =>
-            log.warn(s"Transaction for $p not found, removing")
-            run(delPayout(p.id))
-        }
-    }
-
-    executeAction("""
-                    |alter table payouts
-                    |    drop column active_leases, tx_id, tx_height, confirmed;
-                    |""".stripMargin)
   }
 }
